requirements.txt
streamlit>=1.36.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
